Better organization for action package
