+++ packages/action/src/contract/erc20/ERC20Full.ts
@@ -1,12 +1,12 @@
-import Contract from "../contract/Contract";
+import Contract from "../Contract";
 import TasitContracts from "tasit-contracts";
 const { local } = TasitContracts;
-const { Marketplace } = local;
-const { abi } = Marketplace;
+const { MyERC20Full } = local;
+const { abi } = MyERC20Full;
 
 const abiString = JSON.stringify(abi);
 
-export default class Decentraland extends Contract {
+export default class ERC20Full extends Contract {
   constructor(address: string, account) {
     super(address, abiString, account);
   }
