+++ packages/action/src/contract/erc20/Mana.ts
@@ -1,12 +1,12 @@
-import Contract from "../contract/Contract";
+import Contract from "../Contract";
 import TasitContracts from "tasit-contracts";
 const { local } = TasitContracts;
-const { MyERC721Full } = local;
-const { abi } = MyERC721Full;
+const { MANAToken } = local;
+const { abi } = MANAToken;
 
 const abiString = JSON.stringify(abi);
 
-export default class ERC721Full extends Contract {
+export default class Mana extends Contract {
   constructor(address: string, account) {
     super(address, abiString, account);
   }
