+++ packages/action/src/contract/erc721/ERC721Full.ts
@@ -1,12 +1,12 @@
-import Contract from "../contract/Contract";
+import Contract from "../Contract";
 import TasitContracts from "tasit-contracts";
 const { local } = TasitContracts;
-const { LANDRegistry } = local;
-const { abi } = LANDRegistry;
+const { MyERC721Full } = local;
+const { abi } = MyERC721Full;
 
 const abiString = JSON.stringify(abi);
 
-export default class Land extends Contract {
+export default class ERC721Full extends Contract {
   constructor(address: string, account) {
     super(address, abiString, account);
   }
