+++ packages/action/src/contract/erc721/Estate.ts
@@ -1,12 +1,12 @@
-import Contract from "../contract/Contract";
+import Contract from "../Contract";
 import TasitContracts from "tasit-contracts";
 const { local } = TasitContracts;
-const { MANAToken } = local;
-const { abi } = MANAToken;
+const { EstateRegistry } = local;
+const { abi } = EstateRegistry;
 
 const abiString = JSON.stringify(abi);
 
-export default class Mana extends Contract {
+export default class DecentralandEstate extends Contract {
   constructor(address: string, account) {
     super(address, abiString, account);
   }
