+++ packages/action/src/contract/erc721/Land.ts
@@ -1,12 +1,12 @@
-import Contract from "../contract/Contract";
+import Contract from "../Contract";
 import TasitContracts from "tasit-contracts";
 const { local } = TasitContracts;
-const { EstateRegistry } = local;
-const { abi } = EstateRegistry;
+const { LANDRegistry } = local;
+const { abi } = LANDRegistry;
 
 const abiString = JSON.stringify(abi);
 
-export default class DecentralandEstate extends Contract {
+export default class Land extends Contract {
   constructor(address: string, account) {
     super(address, abiString, account);
   }
