+++ packages/action/src/contract/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Contract";
