+++ packages/action/src/contract/marketplace/Decentraland.ts
@@ -1,12 +1,12 @@
-import Contract from "../contract/Contract";
+import Contract from "../Contract";
 import TasitContracts from "tasit-contracts";
 const { local } = TasitContracts;
-const { MyERC20Full } = local;
-const { abi } = MyERC20Full;
+const { Marketplace } = local;
+const { abi } = Marketplace;
 
 const abiString = JSON.stringify(abi);
 
-export default class ERC20Full extends Contract {
+export default class Decentraland extends Contract {
   constructor(address: string, account) {
     super(address, abiString, account);
   }
